Remove random forest pipeline Add working model registry Add grid search CV
