+++ .docker/mlflow_server/requirements.txt
@@ -1,0 +1,1 @@
+mlflow==2.4.0
