+++ conf/base/mlflow.yml
@@ -17,7 +17,7 @@
 # credentials: mlflow_credentials
 
 server:
-  mlflow_tracking_uri: null # if null, will use mlflow.get_tracking_uri() as a default
+  mlflow_tracking_uri: null
   mlflow_registry_uri: null # if null, mlflow_tracking_uri will be used as mlflow default
   credentials: null  # must be a valid key in credentials.yml which refers to a dict of sensitive mlflow environment variables (password, tokens...). See top of the file.
   request_header_provider: # this is only useful to deal with expiring token, see https://github.com/Galileo-Galilei/kedro-mlflow/issues/357
