--- conf/base/parameters/decision_forest.yml
@@ -1,5 +1,0 @@
-# This is a boilerplate parameters config generated for pipeline 'decision_forest'
-# using Kedro 0.18.9.
-#
-# Documentation for this file format can be found in "Parameters"
-# Link: https://docs.kedro.org/en/0.18.9/kedro_project_setup/configuration.html#parameters
