--- src/asi_projekt/pipelines/decision_forest/README.md
@@ -1,21 +1,0 @@
-# Pipeline decision_forest
-
-> *Note:* This is a `README.md` boilerplate generated using `Kedro 0.18.9`.
-
-## Overview
-
-<!---
-Please describe your modular pipeline here.
--->
-
-## Pipeline inputs
-
-<!---
-The list of pipeline inputs.
--->
-
-## Pipeline outputs
-
-<!---
-The list of pipeline outputs.
--->
