--- src/asi_projekt/pipelines/decision_forest/__init__.py
@@ -1,10 +1,0 @@
-"""
-This is a boilerplate pipeline 'decision_forest'
-generated using Kedro 0.18.9
-"""
-
-from .pipeline import create_pipeline
-
-__all__ = ["create_pipeline"]
-
-__version__ = "0.1"
