--- src/tests/pipelines/decision_forest/test_pipeline.py
@@ -1,9 +1,0 @@
-"""
-This is a boilerplate test file for pipeline 'decision_forest'
-generated using Kedro 0.18.9.
-Please add your pipeline tests here.
-
-Kedro recommends using `pytest` framework, more info about it can be found
-in the official documentation:
-https://docs.pytest.org/en/latest/getting-started.html
-"""
